# src/docs/swagger.yaml
definitions:
...
  github_com_krmsaeed_barber_api_dto.CarModelPropertyResponse:
    properties:
      carModelId:
        type: integer
      id:
        type: integer
      property:
        $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.PropertyResponse'
      value:
        type: string
    type: object
...
  github_com_krmsaeed_barber_api_dto.CityResponse:
    properties:
      country:
        $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CountryResponse'
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_krmsaeed_barber_api_dto.CompanyResponse:
    properties:
      country:
        $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CountryResponse'
      id:
        type: integer
      name:
        type: string
    type: object
  github_com_krmsaeed_barber_api_dto.CountryResponse:
    properties:
      cities:
        items:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CityResponse'
        type: array
      companies:
        items:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CompanyResponse'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_krmsaeed_barber_api_dto.CreateCarModelPropertyRequest:
    properties:
      carModelId:
        type: integer
      propertyId:
        type: integer
      value:
        maxLength: 100
        type: string
    required:
    - carModelId
    - propertyId
    - value
    type: object
...
  github_com_krmsaeed_barber_api_dto.CreateCityRequest:
    properties:
      countryId:
        type: integer
      name:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - countryId
    - name
    type: object
...
  github_com_krmsaeed_barber_api_dto.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description: contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
...
  ? github_com_krmsaeed_barber_api_dto.PagedList-github_com_krmsaeed_barber_api_dto_CarModelPropertyResponse
  : properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CarModelPropertyResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_krmsaeed_barber_api_dto.PagedList-github_com_krmsaeed_barber_api_dto_CityResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CityResponse'
        type: array
      pageNumber:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_krmsaeed_barber_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.Sort'
        type: array
    type: object
...
  github_com_krmsaeed_barber_api_dto.PropertyCategoryResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      properties:
        items:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.PropertyResponse'
        type: array
    type: object
  github_com_krmsaeed_barber_api_dto.PropertyResponse:
    properties:
      category:
        $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.PropertyCategoryResponse'
      dataType:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      unit:
        type: string
    type: object
...
  github_com_krmsaeed_barber_api_dto.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
...
  github_com_krmsaeed_barber_api_dto.UpdateCarModelPropertyRequest:
    properties:
      value:
        maxLength: 100
        type: string
    required:
    - value
    type: object
...
  github_com_krmsaeed_barber_api_dto.UpdateCityRequest:
    properties:
      countryId:
        type: integer
      name:
        maxLength: 20
        minLength: 3
        type: string
    type: object
...
  github_com_krmsaeed_barber_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_validations.ValidationError'
        type: array
    type: object
  github_com_krmsaeed_barber_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  github_com_krmsaeed_barber_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/car-model-properties/:
    post:
      consumes:
      - application/json
      description: Create a CarModelProperty
      parameters:
      - description: Create a CarModelProperty
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CreateCarModelPropertyRequest'
      responses:
        "201":
          description: CarModelProperty response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CarModelPropertyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a CarModelProperty
      tags:
      - CarModelProperties
  /v1/car-model-properties/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a CarModelProperty
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a CarModelProperty
      tags:
      - CarModelProperties
    get:
      consumes:
      - application/json
      description: Get a CarModelProperty
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: CarModelProperty response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CarModelPropertyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a CarModelProperty
      tags:
      - CarModelProperties
    put:
      consumes:
      - application/json
      description: Update a CarModelProperty
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a CarModelProperty
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.UpdateCarModelPropertyRequest'
      responses:
        "200":
          description: CarModelProperty response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CarModelPropertyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a CarModelProperty
      tags:
      - CarModelProperties
  /v1/car-model-properties/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get CarModelProperties
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: CarModelProperty response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.PagedList-github_com_krmsaeed_barber_api_dto_CarModelPropertyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get CarModelProperties
      tags:
      - CarModelProperties
...
  /v1/cities/:
    post:
      consumes:
      - application/json
      description: Create a City
      parameters:
      - description: Create a City
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CreateCityRequest'
      responses:
        "201":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a City
      tags:
      - Cities
  /v1/cities/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a City
      tags:
      - Cities
    get:
      consumes:
      - application/json
      description: Get a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a City
      tags:
      - Cities
    put:
      consumes:
      - application/json
      description: Update a City
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a City
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.UpdateCityRequest'
      responses:
        "200":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a City
      tags:
      - Cities
  /v1/cities/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Cities
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.PaginationInputWithFilter'
      responses:
        "200":
          description: City response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_krmsaeed_barber_api_dto.PagedList-github_com_krmsaeed_barber_api_dto_CityResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_krmsaeed_barber_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Cities
      tags:
      - Cities
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"